Replaced json, fixed variables, and finished embed
I replaced all the lines in the json with more accessible ones so that I can get the info easier.

I fixed some of the variables that weren't being filled in properly. Also fixed two bugs where if a person was unranked, it would break the code, and if a person's account had a space in their name, it would break the code.

Finished setting up the embed so that it actually displays correctly and links to the correct website.